more voronoi

+++ src/main/java/andius/voronoi/nodes/Circle.java
@@ -3,9 +3,9 @@ package andius.voronoi.nodes;
 public final class Circle extends Object {
 
     public Point center;
-    public double radius;
+    public float radius;
 
-    public Circle(double centerX, double centerY, double radius) {
+    public Circle(float centerX, float centerY, float radius) {
         super();
         this.center = new Point(centerX, centerY);
         this.radius = radius;
