+++ src/main/java/andius/voronoi/nodes/Point.java
@@ -1,23 +1,18 @@
 package andius.voronoi.nodes;
 
-/**
- * Point.java
- *
- * @author Connor
- */
 public class Point {
 
-    public static double distance(Point _coord, Point _coord0) {
-        return Math.sqrt((_coord.x - _coord0.x) * (_coord.x - _coord0.x) + (_coord.y - _coord0.y) * (_coord.y - _coord0.y));
+    public static float distance(Point _coord, Point _coord0) {
+        return (float) Math.sqrt((_coord.x - _coord0.x) * (_coord.x - _coord0.x) + (_coord.y - _coord0.y) * (_coord.y - _coord0.y));
     }
     
-    public static double distance(Point _coord, double x, double y) {
-        return Math.sqrt((_coord.x - x) * (_coord.x - x) + (_coord.y - y) * (_coord.y - y));
+    public static float distance(Point _coord, float x, float y) {
+        return (float) Math.sqrt((_coord.x - x) * (_coord.x - x) + (_coord.y - y) * (_coord.y - y));
     }
     
-    public double x, y;
+    public float x, y;
 
-    public Point(double x, double y) {
+    public Point(float x, float y) {
         this.x = x;
         this.y = y;
     }
@@ -27,11 +22,11 @@ public class Point {
         return x + ", " + y;
     }
 
-    public double l2() {
+    public float l2() {
         return x * x + y * y;
     }
 
-    public double length() {
-        return Math.sqrt(x * x + y * y);
+    public float length() {
+        return (float) Math.sqrt(x * x + y * y);
     }
 }
